Перенес все методы в один GlobalExceptionHandler
